Abstraction of handling new tab settings
